Integrate CI workflows, model validation, and stable test execution (#10)
