+++ .github/workflows/ci.yml
@@ -1,0 +1,95 @@
+name: CI
+
+on:
+  push:
+    branches: [ main, workflows-katya ]
+  pull_request:
+    branches: [ main, workflows-katya ]
+
+# One simple image name & tag
+env:
+  IMAGE: ghcr.io/${{ github.repository }}:ci
+
+# Allow pushing to GHCR with the GITHUB_TOKEN
+permissions:
+  contents: read
+  packages: write
+
+jobs:
+  build-image:
+    name: Build CI Docker image
+    runs-on: ubuntu-latest
+
+    steps:
+      - name: Checkout repository
+        uses: actions/checkout@v4
+
+      - name: Log in to GitHub Container Registry
+        uses: docker/login-action@v3
+        with:
+          registry: ghcr.io
+          username: ${{ github.actor }}
+          password: ${{ secrets.GITHUB_TOKEN }}
+
+      - name: Build and push image
+        uses: docker/build-push-action@v6
+        with:
+          context: .
+          push: true
+          tags: ${{ env.IMAGE }}
+
+  lint:
+    name: Lint with ruff (inside container). # Syntax errors, code quality issues, style violations, potential bugs
+    continue-on-error: true
+    runs-on: ubuntu-latest
+    needs: build-image
+
+    steps:
+      - name: Checkout repository
+        uses: actions/checkout@v4
+
+      - name: Log in to GitHub Container Registry
+        uses: docker/login-action@v3
+        with:
+          registry: ghcr.io
+          username: ${{ github.actor }}
+          password: ${{ secrets.GITHUB_TOKEN }}
+
+      - name: Run ruff in container
+        run: |
+          docker run --rm \
+            -e MLFLOW_TRACKING_URI="file:/opt/mlruns" \
+            --workdir /app \
+            ${{ env.IMAGE }} \
+            bash -lc '
+              echo "== Ruff =="
+              uv run ruff check .
+            '
+
+  test:
+    name: Run pytest (inside container)
+    continue-on-error: true
+    runs-on: ubuntu-latest
+    needs: build-image
+
+    steps:
+      - name: Checkout repository
+        uses: actions/checkout@v4
+
+      - name: Log in to GitHub Container Registry
+        uses: docker/login-action@v3
+        with:
+          registry: ghcr.io
+          username: ${{ github.actor }}
+          password: ${{ secrets.GITHUB_TOKEN }}
+
+      - name: Run tests in container
+        run: |
+          docker run --rm \
+            -e MLFLOW_TRACKING_URI="file:/opt/mlruns" \
+            --workdir /app \
+            ${{ env.IMAGE }} \
+            bash -lc '
+              echo "== Pytest =="
+              uv run pytest -vv --maxfail=1
+            '
