+++ .gitignore
@@ -1,3 +1,6 @@
 venv 
 .history
+mlruns/
+__pycache__/
+artifacts/
 /data
