+++ Dockerfile
@@ -1,11 +1,28 @@
-FROM python:3.11-slim
+# Use a slim Python image for speed but still Debian-based for apt packages
+FROM python:3.10-slim
 
+RUN apt-get update && \
+    apt-get install -y --no-install-recommends \
+      build-essential gcc git curl ca-certificates libgomp1 tini \
+    && rm -rf /var/lib/apt/lists/*
+
+# Install uv (Astral)
+RUN curl -LsSf https://astral.sh/uv/install.sh | sh
+ENV PATH="/root/.local/bin:${PATH}"
+
+# Workdir
 WORKDIR /app
 
-COPY requirements.txt .
-
-RUN pip install --no-cache-dir -r requirements.txt
-
+# Copy everything at once
 COPY . .
 
+# Install project dependencies using uv (incl. dev deps for lint/tests)
+RUN uv sync --dev
+
+# Make uv-managed .venv visible as default Python
+ENV VIRTUAL_ENV="/app/.venv"
+ENV PATH="/app/.venv/bin:${PATH}"
+
+# Good container hygiene (tini as entrypoint)
+ENTRYPOINT ["/usr/bin/tini", "--"]
 CMD ["bash"]
