+++ mlops_project/__init__.py
@@ -1,0 +1,1 @@
+# from mlops_project import config  # noqa: F401
