+++ mlops_project/deploy.py
@@ -1,0 +1,35 @@
+import time
+
+from mlflow.tracking import MlflowClient
+
+model_version = 1
+model_name = "lead_model"
+client = MlflowClient()
+
+
+def wait_for_deployment(model_name, model_version, stage="Staging"):
+    status = False
+    while not status:
+        model_version_details = dict(
+            client.get_model_version(name=model_name, version=model_version)
+        )
+        if model_version_details["current_stage"] == stage:
+            status = True
+            break
+        else:
+            time.sleep(2)
+    return status
+
+
+model_version_details = dict(
+    client.get_model_version(name=model_name, version=model_version)
+)
+model_status = True
+if model_version_details["current_stage"] != "Staging":
+    client.transition_model_version_stage(
+        name=model_name,
+        version=model_version,
+        stage="Staging",
+        archive_existing_versions=True,
+    )
+    model_status = wait_for_deployment(model_name, model_version, "Staging")
