+++ model/columns_list.json
@@ -1,0 +1,1 @@
+{"column_names": ["purchases", "time_spent", "n_visits", "customer_group_2", "customer_group_3", "customer_group_4", "customer_group_5", "customer_group_6", "customer_group_7", "customer_group_8", "customer_group_9", "onboarding_True"]}
