+++ pyproject.toml
@@ -17,6 +17,25 @@ classifiers = [
 ]
 requires-python = "~=3.10.0"
 
+dependencies = [
+  "dvc",
+  "loguru",
+  "mkdocs",
+  "pip",
+  "python-dotenv",
+  "ruff",
+  "tqdm",
+  "typer",
+  "pandas==2.2.3",
+  "numpy",
+  "matplotlib",
+  "seaborn",
+  "plotly.express",
+  "scikit-learn==1.5.2",
+  "mlflow==2.18.0",
+  "xgboost==2.1.1",
+  "pytest>=9.0.1",
+]
 
 [tool.ruff]
 line-length = 99
@@ -30,3 +49,6 @@ extend-select = ["I"]  # Add import sorting
 known-first-party = ["mlops_project"]
 force-sort-within-sections = true
 
+[tool.pytest.ini_options]
+testpaths = ["tests"]
+python_files = ["test_*.py"]
