--- source/__init__.py
@@ -1,1 +1,0 @@
-from mlops_project import config  # noqa: F401
