+++ start_container.sh
@@ -1,0 +1,4 @@
+docker run -it --rm \
+  --workdir /app \
+  itu-sdse:latest \
+  bash
